# apps/alpha-web/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "EquitySeries": {
        "additionalProperties": false,
        "properties": {
          "drawdown": {
            "items": {
              "type": "number"
            },
            "title": "Drawdown",
            "type": "array"
          },
          "equity": {
            "items": {
              "type": "number"
            },
            "title": "Equity",
            "type": "array"
          },
          "ts": {
            "items": {
              "type": "number"
            },
            "title": "Ts",
            "type": "array"
          }
        },
        "required": [
          "ts",
          "equity",
          "drawdown"
        ],
        "title": "EquitySeries",
        "type": "object"
      },
      "ForecastOrigins": {
        "additionalProperties": false,
        "properties": {
          "cover50": {
            "items": {
              "type": "boolean"
            },
            "title": "Cover50",
            "type": "array"
          },
          "cover80": {
            "items": {
              "type": "boolean"
            },
            "title": "Cover80",
            "type": "array"
          },
          "cover90": {
            "items": {
              "type": "boolean"
            },
            "title": "Cover90",
            "type": "array"
          },
          "crps": {
            "items": {
              "type": "number"
            },
            "title": "Crps",
            "type": "array"
          },
          "crps_bootstrap": {
            "items": {
              "type": "number"
            },
            "title": "Crps Bootstrap",
            "type": "array"
          },
          "crps_rw": {
            "items": {
              "type": "number"
            },
            "title": "Crps Rw",
            "type": "array"
          },
          "hit": {
            "items": {
              "type": "boolean"
            },
            "title": "Hit",
            "type": "array"
          },
          "median_end_return": {
            "items": {
              "type": "number"
            },
            "title": "Median End Return",
            "type": "array"
          },
          "origin_ts": {
            "items": {
              "type": "number"
            },
            "title": "Origin Ts",
            "type": "array"
          },
          "pre_cutoff": {
            "items": {
              "type": "boolean"
            },
            "title": "Pre Cutoff",
            "type": "array"
          },
          "realized_end_return": {
            "items": {
              "type": "number"
            },
            "title": "Realized End Return",
            "type": "array"
          }
        },
        "required": [
          "origin_ts",
          "pre_cutoff",
          "crps",
          "crps_rw",
          "crps_bootstrap",
          "realized_end_return",
          "median_end_return",
          "hit",
          "cover50",
          "cover80",
          "cover90"
        ],
        "title": "ForecastOrigins",
        "type": "object"
      },
      "ForecastPath": {
        "additionalProperties": false,
        "properties": {
          "closes": {
            "items": {
              "type": "number"
            },
            "title": "Closes",
            "type": "array"
          },
          "sample": {
            "title": "Sample",
            "type": "integer"
          }
        },
        "required": [
          "sample",
          "closes"
        ],
        "title": "ForecastPath",
        "type": "object"
      },
      "ForecastPaths": {
        "additionalProperties": false,
        "properties": {
          "samples": {
            "items": {
              "$ref": "#/components/schemas/ForecastPath"
            },
            "title": "Samples",
            "type": "array"
          },
          "ts": {
            "items": {
              "type": "number"
            },
            "title": "Ts",
            "type": "array"
          }
        },
        "required": [
          "samples",
          "ts"
        ],
        "title": "ForecastPaths",
        "type": "object"
      },
      "ForecastSeries": {
        "additionalProperties": false,
        "properties": {
          "forecast": {
            "items": {
              "type": "number"
            },
            "title": "Forecast",
            "type": "array"
          },
          "forecast_ts": {
            "items": {
              "type": "number"
            },
            "title": "Forecast Ts",
            "type": "array"
          },
          "history": {
            "items": {
              "type": "number"
            },
            "title": "History",
            "type": "array"
          },
          "history_ts": {
            "items": {
              "type": "number"
            },
            "title": "History Ts",
            "type": "array"
          },
          "mean": {
            "items": {
              "type": "number"
            },
            "title": "Mean",
            "type": "array"
          },
          "q05": {
            "items": {
              "type": "number"
            },
            "title": "Q05",
            "type": "array"
          },
          "q25": {
            "items": {
              "type": "number"
            },
            "title": "Q25",
            "type": "array"
          },
          "q75": {
            "items": {
              "type": "number"
            },
            "title": "Q75",
            "type": "array"
          },
          "q95": {
            "items": {
              "type": "number"
            },
            "title": "Q95",
            "type": "array"
          }
        },
        "required": [
          "history_ts",
          "history",
          "forecast_ts",
          "forecast",
          "q05",
          "q25",
          "q75",
          "q95",
          "mean"
        ],
        "title": "ForecastSeries",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NullTier": {
        "additionalProperties": false,
        "properties": {
          "statistics": {
            "items": {
              "type": "number"
            },
            "title": "Statistics",
            "type": "array"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "tier",
          "statistics"
        ],
        "title": "NullTier",
        "type": "object"
      },
      "NullTiers": {
        "additionalProperties": false,
        "properties": {
          "tiers": {
            "items": {
              "$ref": "#/components/schemas/NullTier"
            },
            "title": "Tiers",
            "type": "array"
          }
        },
        "required": [
          "tiers"
        ],
        "title": "NullTiers",
        "type": "object"
      },
      "OptimTrial": {
        "additionalProperties": false,
        "properties": {
          "returns": {
            "items": {
              "type": "number"
            },
            "title": "Returns",
            "type": "array"
          },
          "trial": {
            "title": "Trial",
            "type": "integer"
          }
        },
        "required": [
          "trial",
          "returns"
        ],
        "title": "OptimTrial",
        "type": "object"
      },
      "OptimTrials": {
        "additionalProperties": false,
        "properties": {
          "trials": {
            "items": {
              "$ref": "#/components/schemas/OptimTrial"
            },
            "title": "Trials",
            "type": "array"
          }
        },
        "required": [
          "trials"
        ],
        "title": "OptimTrials",
        "type": "object"
      },
...
      "PropfirmPathColumns": {
        "additionalProperties": false,
        "properties": {
          "busted": {
            "items": {
              "type": "boolean"
            },
            "title": "Busted",
            "type": "array"
          },
          "days_to_pass": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Days To Pass",
            "type": "array"
          },
          "passed": {
            "items": {
              "type": "boolean"
            },
            "title": "Passed",
            "type": "array"
          },
          "payout": {
            "items": {
              "type": "number"
            },
            "title": "Payout",
            "type": "array"
          }
        },
        "required": [
          "passed",
          "busted",
          "days_to_pass",
          "payout"
        ],
        "title": "PropfirmPathColumns",
        "type": "object"
      },
      "PropfirmPaths": {
        "additionalProperties": false,
        "properties": {
          "paths": {
            "$ref": "#/components/schemas/PropfirmPathColumns"
          }
        },
        "required": [
          "paths"
        ],
        "title": "PropfirmPaths",
        "type": "object"
      },
...
      "RunDetail": {
        "additionalProperties": false,
        "properties": {
          "has_equity": {
            "title": "Has Equity",
            "type": "boolean"
          },
          "has_forecast": {
            "title": "Has Forecast",
            "type": "boolean"
          },
          "has_forecast_paths": {
            "title": "Has Forecast Paths",
            "type": "boolean"
          },
          "has_nulls": {
            "title": "Has Nulls",
            "type": "boolean"
          },
          "has_origins": {
            "title": "Has Origins",
            "type": "boolean"
          },
          "has_propfirm_paths": {
            "title": "Has Propfirm Paths",
            "type": "boolean"
          },
          "has_tearsheet": {
            "title": "Has Tearsheet",
            "type": "boolean"
          },
          "has_trades": {
            "title": "Has Trades",
            "type": "boolean"
          },
          "has_trials": {
            "title": "Has Trials",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "manifest": {
            "additionalProperties": true,
            "title": "Manifest",
            "type": "object"
          },
          "mtime": {
            "title": "Mtime",
            "type": "number"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "kind",
          "mtime",
          "manifest",
          "has_equity",
          "has_trades",
          "has_tearsheet",
          "has_forecast",
          "has_nulls",
          "has_trials",
          "has_forecast_paths",
          "has_propfirm_paths",
          "has_origins"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items"
        ],
        "title": "RunList",
        "type": "object"
      },
      "RunListItem": {
        "additionalProperties": false,
        "properties": {
          "command": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "mtime": {
            "title": "Mtime",
            "type": "number"
          },
          "passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passed"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "symbols": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbols"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          }
        },
        "required": [
          "run_id",
          "kind",
          "command",
          "label",
          "symbol",
          "symbols",
          "passed",
          "verdict",
          "mtime"
        ],
        "title": "RunListItem",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Project ALPHA \u2014 Workstation",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs": {
      "get": {
        "description": "Filtered, newest-first, paginated run index for the run browser.",
        "operationId": "list_runs_api_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "symbol",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          {
            "in": "query",
            "name": "verdict",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "in": "query",
            "name": "passed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Passed"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "description": "Full manifest + kind/mtime + artifact-presence flags.",
        "operationId": "run_detail_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Detail",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/equity": {
      "get": {
        "description": "The run's equity curve as ``{ts, equity, drawdown}`` (empty lists when it wrote none).",
        "operationId": "run_equity_api_runs__run_id__equity_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquitySeries"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Equity",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/forecast": {
      "get": {
        "description": "A forecast run's history + forecast series (+ quantile bands). 404 otherwise.",
        "operationId": "run_forecast_api_runs__run_id__forecast_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastSeries"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Forecast",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/forecast/paths": {
      "get": {
        "description": "The first ``n`` (clamped to 40) sampled close paths of a forecast run. 404 otherwise.",
        "operationId": "run_forecast_paths_api_runs__run_id__forecast_paths_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "n",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 40,
              "minimum": 1,
              "title": "N",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastPaths"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Forecast Paths",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/nulls": {
      "get": {
        "description": "A gauntlet run's raw per-tier null distributions. 404 when it wrote none.",
        "operationId": "run_nulls_api_runs__run_id__nulls_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NullTiers"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Nulls",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/origins": {
      "get": {
        "description": "A forecast-eval run's per-origin skill scores, columnar. 404 when it wrote none.",
        "operationId": "run_origins_api_runs__run_id__origins_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastOrigins"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Origins",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/propfirm-paths": {
      "get": {
        "description": "A prop-firm run's per-path Monte-Carlo outcomes, columnar. 404 when it wrote none.",
        "operationId": "run_propfirm_paths_api_runs__run_id__propfirm_paths_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropfirmPaths"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Propfirm Paths",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/tearsheet": {
      "get": {
        "description": "The run's rendered quantstats tear sheet (embedded in Run Detail via an iframe).",
        "operationId": "run_tearsheet_api_runs__run_id__tearsheet_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Tearsheet",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/trades": {
      "get": {
        "description": "The run's trade-log rows (``[]`` when it wrote none).",
        "operationId": "run_trades_api_runs__run_id__trades_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Run Trades Api Runs  Run Id  Trades Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Trades",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/runs/{run_id}/trials": {
      "get": {
        "description": "A sweep's per-config OOS return streams. 404 when it wrote none.",
        "operationId": "run_trials_api_runs__run_id__trials_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimTrials"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Trials",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}